reorganize code; implement terminal resizing, add startup command param, etc
